rephrase statements about check-in procedure
I avoid using the term 'registration' for both ticket buying and
on-site identity confirmation. I currently put the latter as 'check in'
but I am not sure whether reg team will agree on this decision.

Also list style on mutliple id verification is changed, to enumerated
list is a bit confusing. Work on #45, #50